requirements.txt
fastapi>=0.104.1
pillow>=10.1.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
uvicorn[standard]>=0.24.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
greenlet>=3.2.0
python-multipart>=0.0.6

